+++ packages/gatsby-source-nacelle/src/gql-fragments/Product.graphql
