+++ packages/gatsby-source-nacelle/src/gql-fragments/Space.graphql
