+++ packages/gatsby-source-nacelle/src/source-nodes/index.js
@@ -1,0 +1,4 @@
+// Node sourcing utilities that can be called in the exports.sourceNodes of `gatsby-node.js`
+
+exports.sourceContentfulPreviewNodes = require('./contentful-preview');
+exports.sourceNacelleNodes = require('./nacelle');
