+++ packages/gatsby-source-nacelle/src/utils/getNacelleVariables.js
