+++ packages/gatsby-source-nacelle/src/utils/index.js
@@ -1,0 +1,2 @@
+exports.getNacelleVariables = require('./getNacelleVariables');
+exports.replaceKeys = require('./replaceKeys');
