+++ packages/gatsby-source-nacelle/src/utils/replaceKeys.js
