--- packages/gatsby-source-nacelle/utils/index.js
@@ -1,2 +1,0 @@
-exports.envVarHandler = require('./envVarHandler');
-exports.replaceKeys = require('./replaceKeys');
